requirements.txt
google-generativeai
prometheus-fastapi-instrumentator
fastapi
uvicorn
streamlit
openai
anthropic
sympy
pyyaml
python-dotenv
locust
requests
